repo: move rebase_descendants_with_options_return_map() to tests
